Add assertions for classic transaction management within context
For pure psycopg2.extensions.connection instances this is not possible.
(It is not possible to monkey-patch types in extension modules.)

For connection types derived in Python, patch `cxn.commit()` and
`cxn.rollback()` to raise helpful assertions.

Attempting to set `cxn.autocommit = True` already raises (albeit a
rather obscure psycopg2.InternalError('no such savepoint X')) and
patching is involved and risky (because it is a property) so I've left
`autocommit` unpatched for now.