Ensure patching commit/rollback does not introduce instance-level attrs
In the usual scenario, commit() and rollback() are methods defined on
the connection type, not the instance. Naive patching and restore which
simplifies down to `cxn.commit = cxn.commit` is not a no-op, because it
will take a method defined on the connection type and turn it in to a
bound instance method and add it to the instance dict of the connection.

The correct approach is to take note if the method is defined locally
before patching, and if not, undo the patch by deleting from the
instance dict, rather than assigning to it.

Fixes #11