Add tests for connections with autocommit off but not yet in transaction
Specifically, assert that a connection with autocommit off that has not
executed any statements (and therefore has not started a transaction
yet) behaves the same as a connection with autocommit on.

This means that library code receiving a connection and using it with
Transaction() is completely agnostic to connection's autocommit state.